refactor: pull out and tidy http/s proxy
